pyproject.toml
[project]
name = "streamflow-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.3",
    "geoalchemy2>=0.15.2",
    "greenlet>=3.1.1",
    "httpx>=0.27.2",
    "polars>=1.10.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
